Justified classes to new schema

+++ src/main/java/oop/CourseWork/model/check/CheckRepository.java
@@ -2,5 +2,5 @@ package oop.CourseWork.model.check;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CheckRepository extends JpaRepository<Check, Integer> {
+public interface CheckRepository extends JpaRepository<Check, Long> {
 }
