+++ src/main/java/oop/CourseWork/model/productBase/ProductBaseRepository.java
@@ -2,5 +2,5 @@ package oop.CourseWork.model.productBase;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductBaseRepository extends JpaRepository<ProductBase, Integer> {
+public interface ProductBaseRepository extends JpaRepository<ProductBase, Long> {
 }
